Add Quiet (-q|--quiet) to suppress Ctrl+D banner
Previously, all recordings would display, "Press Ctrl+D to end
recording".  This is helpful (& even desirable) on first usage.
However, it is annoying to have the banner in all the recordings.

This adds the ability to add -q|--quiet, which suppresses the banner
making cleaner recordings.  (An empty string is outputted as the
application needs there to be something to begin recording)

Signed-off-by: Dave Walker (Daviey) <email@daviey.com>